pyproject.toml
[build-system]
requires = [
    "setuptools >= 65",
    "wheel >= 0.38"
]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = ["pyenv*", "tests*"]

[project]
name = "chatcmd"
version = "2.2.1"
description = "Open-source AI CLI for command lookup, SQL generation, and developer tools with multi-model support (OpenAI, Anthropic, Google, Cohere, Ollama)."
authors = [
    { name = "Naif Alshaye", email = "naif@naif.io" }
]
license = {text = "MIT License"}
readme = "README.md"
requires-python = ">=3.8.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    'Intended Audience :: System Administrators',
    'Intended Audience :: Developers',
    'Intended Audience :: End Users/Desktop',
    "Natural Language :: English",
    "Topic :: Terminals",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: Implementation",
]
keywords = [
    "cli", "command", "command-line", "commandline", "terminal", "development","linux", "ai", "artificial-intelligence", "chatcmd", "chatgpt", "gpt-4", "gpt-3.5-turbo", "gpt-3.5-turbo", "claude", "gemini", "ollama", "lookup", "SQL", "regex", "base64", "open source","MIT"
]
dependencies = [
    "docopt >=0.6.2,<0.7.0",
    "openai >=1.0.0,<2.0.0",
    "anthropic >=0.7.0,<1.0.0",
    "google-genai >=1.0.0",
    "cohere >=4.0.0,<5.0.0",
    "pyperclip >=1.8.2,<2.0.0",
    "requests >=2.31.0,<3.0.0",
    "fake_useragent >=1.3.0,<2.0.0",
    "cryptography >=41.0.0,<43.0.0",
    "keyring >=23.0.0,<25.0.0"
]

[project.urls]
"Homepage" = "https://github.com/naifalshaye/chatcmd"
"Documentation" = "https://github.com/naifalshaye/chatcmd/blob/master/README.md"
"Repository" = "https://github.com/naifalshaye/chatcmd.git"
"Bug Tracker" = "https://github.com/naifalshaye/chatcmd/issues"

[project.scripts]
chatcmd = "chatcmd.chatcmd_init:main"